requirements.txt
Django==5.0.9
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.4.0
djangorestframework==3.15.2
django-cors-headers==4.4.0
celery==5.4.0
redis==5.0.8
gunicorn==22.0.0
whitenoise==6.7.0
django-storages==1.14.4
boto3==1.34.162
cryptography==43.0.0
PyJWT==2.9.0
qrcode[pil]==7.4.2
reportlab==4.2.2
django-import-export==4.1.1
